+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - do not create a group if the sender includes self in the To field
 
 ### Fixed
+- Display message error even for messages that are not OutFailed #5498
 - fix app picker sometimes incorrectly showing "Offline"
 - allow using the app picker even when offline, as long as the app store data is cached
 - if adding an app from the app picker fails, show an error dialog